+++ config/routes.rb
@@ -166,10 +166,7 @@ Rails.application.routes.draw do
       member do
         get 'get_annotations'
         get 'add_extra_marks'
-        get 'download'
         post 'download'
-        get 'download_zip'
-        delete 'cancel_remark_request'
         post 'add_extra_mark'
         delete 'delete_grace_period_deduction'
         get 'next_grouping'
@@ -188,7 +185,6 @@ Rails.application.routes.draw do
         get 'get_test_runs_instructors'
         get 'get_test_runs_instructors_released'
         get 'view_token_check'
-        patch 'update_remark_request'
       end
     end
 
@@ -322,6 +318,12 @@ Rails.application.routes.draw do
           post 'repo_browser'
           post 'manually_collect_and_begin_grading'
           get 'downloads'
+        end
+        member do
+          get 'download_file'
+          get 'download_file_zip'
+          delete 'cancel_remark_request'
+          patch 'update_remark_request'
         end
       end