+++ spec/controllers/admin/course_controller_spec.rb
@@ -268,7 +268,7 @@ describe Admin::CoursesController do
       context 'there is no autotest_setting set' do
         it 'should return unprocessable_entity' do
           subject
-          assert_response :unprocessable_entity
+          expect(response).to have_http_status(:unprocessable_entity)
         end
       end
       context 'there is an autotest_setting' do
@@ -305,7 +305,7 @@ describe Admin::CoursesController do
       context 'there is no autotest_setting set' do
         it 'should return unprocessable_entity' do
           subject
-          assert_response :unprocessable_entity
+          expect(response).to have_http_status(:unprocessable_entity)
         end
       end
       context 'there is an autotest_setting' do