+++ Changelog.md
@@ -1,6 +1,9 @@
+
 # Changelog
 
 ## [unreleased]
+- Move submission-specific results/ routes to be under submissions/ (#6434)
+- Fix bug where in some circumstances the wrong result would be displayed to students (#6465)
 - Add option to allow Cross-Origin Resource Sharing (CORS) from JupyterHub (#6442)
 
 ## [v2.2.2]